moved phone permissions file to utilities package
